+++ src/dspf-edit.model.ts
@@ -24,6 +24,7 @@ export type DdsElement =
 export interface DdsAttribute {
   kind: 'attribute';
   lineIndex: number;
+  lastLineIndex?: number;
   value: string;
   attribute?: string;
   indicators?: DdsIndicator[];
@@ -198,6 +199,8 @@ export interface FieldInfo {
   length: number;
   attributes: AttributeWithIndicators[];
   indicators?: DdsIndicator[];
+  lineIndex: number;
+  lastLineIndex: number;
 };
 
 /** Simplified constant info */
@@ -208,12 +211,16 @@ export interface ConstantInfo {
   col: number;
   length: number;
   attributes: AttributeWithIndicators[];
+  lineIndex: number;
+  lastLineIndex: number;
 };
 
 /*** Attribute with indicadors info */
 export interface AttributeWithIndicators {
   value: string;
   indicators?: DdsIndicator[];
+  lineIndex: number;
+  lastLineIndex: number;
 };
 
 /** Record container for fields & constants */